+++ src/components/music/card/MusicRoomCard.tsx
@@ -2,8 +2,8 @@ import Image from 'next/image';
 import { BsPlayFill } from 'react-icons/bs';
 import { AiFillEye } from 'react-icons/ai';
 import Link from 'next/link';
-import { Room } from '@/app/service/room';
 import MusicDeleteModal from '../MusicDeleteModal';
+import { Room } from '@/types/room';
 
 type Props = {
 	musicRoom: Room;
